+++ common/src/main/java/earth/terrarium/athena/impl/client/models/ctm/FourtySevenSliceCtmProvider.java
@@ -2,6 +2,7 @@ package earth.terrarium.athena.impl.client.models.ctm;
 
 import com.google.common.base.Preconditions;
 import com.mojang.serialization.Codec;
+import com.mojang.serialization.DynamicOps;
 import com.mojang.serialization.DataResult;
 import earth.terrarium.athena.api.client.models.AthenaQuad;
 import earth.terrarium.athena.api.client.utils.CtmState;
@@ -65,6 +66,11 @@ public record FourtySevenSliceCtmProvider(
     }
 
     public record Type(String texture) implements CtmProvider.Type {
+        @Override
+        public <T> DataResult<T> encode(DynamicOps<T> ops, T prefix) {
+            return CODEC.encode(this, ops, prefix);
+        }
+
         @Override
         public CtmProvider build(MaterialStorage materials) {
             int[] sprites = new int[47];