+++ common/src/main/java/earth/terrarium/athena/impl/client/models/ctm/SingleSpriteCtmProvider.java
@@ -1,6 +1,8 @@
 package earth.terrarium.athena.impl.client.models.ctm;
 
 import com.mojang.serialization.Codec;
+import com.mojang.serialization.DynamicOps;
+import com.mojang.serialization.DataResult;
 import earth.terrarium.athena.api.client.models.AthenaQuad;
 import earth.terrarium.athena.api.client.utils.CtmState;
 import earth.terrarium.athena.impl.client.models.materials.MaterialStorage;
@@ -19,6 +21,11 @@ public record SingleSpriteCtmProvider(
     }
 
     public record Type(Material material) implements CtmProvider.Type {
+        @Override
+        public <T> DataResult<T> encode(DynamicOps<T> ops, T prefix) {
+            return CODEC.encode(this, ops, prefix);
+        }
+
         @Override
         public CtmProvider build(MaterialStorage materials) {
             return new SingleSpriteCtmProvider(materials.put(material));